requirements.txt
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
click==8.1.7
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.4
